+++ src/core/variableconv_mc.jl
@@ -33,310 +33,313 @@ end
 
 "variable: `pconv_ac[j]` for `j` in `convdc`"
 function variable_converter_active_power(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
-    pc = _PM.var(pm, nw)[:pconv_ac] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_pconv_ac_$(i)",
-        start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:pconv_ac] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_pconv_ac_$(i)"
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(pc[c], convdc["Pacmin"])
-            JuMP.set_upper_bound.(pc[c], convdc["Pacmax"])
+    for (i, convdc) in _PM.ref(pm, nw, :convdc)
+        JuMP.set_start_value.(vars[i], comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", first(conductors[i]), 1.0))
+        if bounded
+            JuMP.set_lower_bound.(vars[i], convdc["Pacmin"][first(conductors[i])])
+            JuMP.set_upper_bound.(vars[i], convdc["Pacmax"][first(conductors[i])])
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :pconv, _PM.ids(pm, nw, :convdc), pc)
+    report && sol_component_value_status(pm, nw, :convdc, :pconv, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `qconv_ac[j]` for `j` in `convdc`"
 function variable_converter_reactive_power(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
-    qc = _PM.var(pm, nw)[:qconv_ac] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_qconv_ac_$(i)",
-        start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "Q_g", c, 1.0)
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:qconv_ac] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_qconv_ac_$(i)"
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(qc[c], convdc["Qacmin"])
-            JuMP.set_upper_bound.(qc[c], convdc["Qacmax"])
+    for (i, convdc) in _PM.ref(pm, nw, :convdc)
+        JuMP.set_start_value.(vars[i], comp_start_value(_PM.ref(pm, nw, :convdc, i), "Q_g", first(conductors[i]), 1.0))
+        if bounded
+            JuMP.set_lower_bound.(vars[i], convdc["Qacmin"][first(conductors[i])])
+            JuMP.set_upper_bound.(vars[i], convdc["Qacmax"][first(conductors[i])])
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :qconv, _PM.ids(pm, nw, :convdc), qc)
+    report && sol_component_value_status(pm, nw, :convdc, :qconv, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `pconv_grid_ac_to[j]` for `j` in `convdc`"
 function variable_conv_transformer_active_power_to(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 2
-
-    ptfto = _PM.var(pm, nw)[:pconv_tf_to] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_pconv_tf_to_$(i)",
-        start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:pconv_tf_to] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_pconv_tf_to_$(i)"
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(ptfto[c], -convdc["Pacrated"] * bigM)
-            JuMP.set_upper_bound.(ptfto[c], convdc["Pacrated"] * bigM)
+    for (i, convdc) in _PM.ref(pm, nw, :convdc)
+        JuMP.set_start_value.(vars[i], comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", first(conductors[i]), 1.0))
+        if bounded
+            JuMP.set_lower_bound.(vars[i], -convdc["Pacrated"][first(conductors[i])] * bigM)
+            JuMP.set_upper_bound.(vars[i], convdc["Pacrated"][first(conductors[i])] * bigM)
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :ptf_to, _PM.ids(pm, nw, :convdc), ptfto)
+    report && sol_component_value_status(pm, nw, :convdc, :ptf_to, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `qconv_grid_ac_to[j]` for `j` in `convdc`"
 function variable_conv_transformer_reactive_power_to(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 2
-    qtfto = _PM.var(pm, nw)[:qconv_tf_to] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_qconv_tf_to__$(i)",
-        start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "Q_g", c, 1.0)
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:qconv_tf_to] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_qconv_tf_to__$(i)"
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(qtfto[c], -convdc["Qacrated"] * bigM)
-            JuMP.set_upper_bound.(qtfto[c], convdc["Qacrated"] * bigM)
-
+    for (i, convdc) in _PM.ref(pm, nw, :convdc)
+        JuMP.set_start_value.(vars[i], comp_start_value(_PM.ref(pm, nw, :convdc, i), "Q_g", first(conductors[i]), 1.0))
+        if bounded
+            JuMP.set_lower_bound.(vars[i], -convdc["Qacrated"][first(conductors[i])] * bigM)
+            JuMP.set_upper_bound.(vars[i], convdc["Qacrated"][first(conductors[i])] * bigM)
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :qtf_to, _PM.ids(pm, nw, :convdc), qtfto)
+    report && sol_component_value_status(pm, nw, :convdc, :qtf_to, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `pconv_pr_from[j]` for `j` in `convdc`"
 function variable_conv_reactor_active_power_from(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 2
-    pprfr = _PM.var(pm, nw)[:pconv_pr_fr] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_pconv_pr_fr_$(i)",
-        start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:pconv_pr_fr] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_pconv_pr_fr_$(i)"
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(pprfr[c], -convdc["Pacrated"] * bigM)
-            JuMP.set_upper_bound.(pprfr[c], convdc["Pacrated"] * bigM)
+    for (i, convdc) in _PM.ref(pm, nw, :convdc)
+        JuMP.set_start_value.(vars[i], comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", first(conductors[i]), 1.0))
+        if bounded
+            JuMP.set_lower_bound.(vars[i], -convdc["Pacrated"][first(conductors[i])] * bigM)
+            JuMP.set_upper_bound.(vars[i], convdc["Pacrated"][first(conductors[i])] * bigM)
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :ppr_fr, _PM.ids(pm, nw, :convdc), pprfr)
+    report && sol_component_value_status(pm, nw, :convdc, :ppr_fr, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `qconv_pr_from[j]` for `j` in `convdc`"
 function variable_conv_reactor_reactive_power_from(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 2
-    qprfr = _PM.var(pm, nw)[:qconv_pr_fr] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_qconv_pr_fr_$(i)",
-        start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "Q_g", c, 1.0)
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:qconv_pr_fr] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_qconv_pr_fr_$(i)"
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(qprfr[c], -convdc["Qacrated"] * bigM)
-            JuMP.set_upper_bound.(qprfr[c], convdc["Qacrated"] * bigM)
+    for (i, convdc) in _PM.ref(pm, nw, :convdc)
+        JuMP.set_start_value.(vars[i], comp_start_value(_PM.ref(pm, nw, :convdc, i), "Q_g", first(conductors[i]), 1.0))
+        if bounded
+            JuMP.set_lower_bound.(vars[i], -convdc["Qacrated"][first(conductors[i])] * bigM)
+            JuMP.set_upper_bound.(vars[i], convdc["Qacrated"][first(conductors[i])] * bigM)
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :qpr_fr, _PM.ids(pm, nw, :convdc), qprfr)
+    report && sol_component_value_status(pm, nw, :convdc, :qpr_fr, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `pconv_grid_ac[j]` for `j` in `convdc`"
 function variable_converter_to_grid_active_power(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 2
-    ptffr = _PM.var(pm, nw)[:pconv_tf_fr] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_pconv_tf_fr_$(i)",
-        start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:pconv_tf_fr] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_pconv_tf_fr_$(i)"
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(ptffr[c], -convdc["Pacrated"] * bigM)
-            JuMP.set_upper_bound.(ptffr[c], convdc["Pacrated"] * bigM)
+    for (i, convdc) in _PM.ref(pm, nw, :convdc)
+        JuMP.set_start_value.(vars[i], comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", first(conductors[i]), 1.0))
+        if bounded
+            JuMP.set_lower_bound.(vars[i], -convdc["Pacrated"][first(conductors[i])] * bigM)
+            JuMP.set_upper_bound.(vars[i], convdc["Pacrated"][first(conductors[i])] * bigM)
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :pgrid, _PM.ids(pm, nw, :convdc), ptffr)
+    report && sol_component_value_status(pm, nw, :convdc, :pgrid, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `qconv_grid_ac[j]` for `j` in `convdc`"
 function variable_converter_to_grid_reactive_power(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 2
-    qtffr = _PM.var(pm, nw)[:qconv_tf_fr] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_qconv_tf_fr_$(i)",
-        start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "Q_g", c, 1.0)
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:qconv_tf_fr] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_qconv_tf_fr_$(i)"
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(qtffr[c], -convdc["Qacrated"] * bigM)
-            JuMP.set_upper_bound.(qtffr[c], convdc["Qacrated"] * bigM)
+    for (i, convdc) in _PM.ref(pm, nw, :convdc)
+        JuMP.set_start_value.(vars[i], comp_start_value(_PM.ref(pm, nw, :convdc, i), "Q_g", first(conductors[i]), 1.0))
+        if bounded
+            JuMP.set_lower_bound.(vars[i], -convdc["Qacrated"][first(conductors[i])] * bigM)
+            JuMP.set_upper_bound.(vars[i], convdc["Qacrated"][first(conductors[i])] * bigM)
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :qgrid, _PM.ids(pm, nw, :convdc), qtffr)
+    report && sol_component_value_status(pm, nw, :convdc, :qgrid, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 function variable_dcside_current(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
-    icdc = _PM.var(pm, nw)[:iconv_dc] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]+1], base_name = "$(nw)_iconv_dc_$(i)",
+    conductors = _PM.ref(pm, nw, :convs_dc_cond)
+    vars = _PM.var(pm, nw)[:iconv_dc] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_iconv_dc_$(i)",
         start = 1.0
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for i in _PM.ids(pm, nw, :convdc)
-            conv_cond = _PM.ref(pm, nw, :convdc)[i]["conductors"]
-            for c in 1:conv_cond
-                JuMP.set_lower_bound.(icdc[i][c], -
-                _PM.ref(pm, nw, :convdc)[i]["Imax"][c])
-                JuMP.set_upper_bound.(icdc[i][c], _PM.ref(pm, nw, :convdc)[i]["Imax"][c])
-            end
-            JuMP.set_lower_bound.(icdc[i][conv_cond+1], -_PM.ref(pm, nw, :convdc)[i]["Imax"][1])
-            JuMP.set_upper_bound.(icdc[i][conv_cond+1], _PM.ref(pm, nw, :convdc)[i]["Imax"][1])
+    if bounded # The current rating of the metallic return converter terminal is set equal to the rating of the first converter pole
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            active_pole = first(conductors[i])[1:end-1]
+            JuMP.set_lower_bound.(vars[i], -vcat(convdc["Imax"][active_pole], first(convdc["Imax"])))
+            JuMP.set_upper_bound.(vars[i], vcat(convdc["Imax"][active_pole], first(convdc["Imax"])))
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :iconv_dc, _PM.ids(pm, nw, :convdc), icdc)
-
+    report && sol_component_value_status(pm, nw, :convdc, :iconv_dc, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 function variable_dcside_current_ground(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
-    icdcg = _PM.var(pm, nw)[:iconv_dcg] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_iconv_dcg_$(i)",
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:iconv_dcg] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_iconv_dcg_$(i)",
         start = 1.0
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
     if bounded
-        for i in _PM.ids(pm, nw, :convdc)
-            conv_cond = _PM.ref(pm, nw, :convdc)[i]["conductors"]
-            for c in 1:conv_cond
-                JuMP.set_lower_bound.(icdcg[i][c], -_PM.ref(pm, nw, :convdc)[i]["Imax"][c])
-                JuMP.set_upper_bound.(icdcg[i][c], _PM.ref(pm, nw, :convdc)[i]["Imax"][c])
-            end
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound.(vars[i], -convdc["Imax"][first(conductors[i])])
+            JuMP.set_upper_bound.(vars[i], convdc["Imax"][first(conductors[i])])
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :iconv_dcg, _PM.ids(pm, nw, :convdc), icdcg)
+    report && sol_component_value_status(pm, nw, :convdc, :iconv_dcg, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 function variable_dcside_current_grounding_shunt(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 1.2 # to account for losses, maximum losses to be derived
-    icdcg_shunt = _PM.var(pm, nw)[:iconv_dcg_shunt] = Dict(i => JuMP.@variable(pm.model,
+    vars = _PM.var(pm, nw)[:iconv_dcg_shunt] = Dict(i => JuMP.@variable(pm.model,
         base_name = "$(nw)_iconv_dcg_shunt_$(i)",
         start = 1.0
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
     if bounded
-        for i in _PM.ids(pm, nw, :convdc)
-            JuMP.set_lower_bound.(icdcg_shunt[i], -_PM.ref(pm, nw, :convdc)[i]["Imax"][1] * 0.1 * bigM * _PM.ref(pm, nw, :convdc)[i]["ground_type"])
-            JuMP.set_upper_bound.(icdcg_shunt[i], _PM.ref(pm, nw, :convdc)[i]["Imax"][1] * 0.1 * bigM * _PM.ref(pm, nw, :convdc)[i]["ground_type"])
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound(vars[i], -convdc["Imax"][1] * 0.1 * bigM * convdc["ground_type"])
+            JuMP.set_upper_bound(vars[i], convdc["Imax"][1] * 0.1 * bigM * convdc["ground_type"])
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :iconv_dcg_shunt, _PM.ids(pm, nw, :convdc), icdcg_shunt)
+    report && _PM.sol_component_value(pm, nw, :convdc, :iconv_dcg_shunt, _PM.ids(pm, nw, :convdc), vars)
 end
 
 "variable: `pconv_dc[j]` for `j` in `convdc`"
 function variable_dcside_power(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 1.2 # to account for losses, maximum losses to be derived
-    pcdc = _PM.var(pm, nw)[:pconv_dc] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]+1], base_name = "$(nw)_pconv_dc_$(i)",
+    conductors = _PM.ref(pm, nw, :convs_dc_cond)
+    vars = _PM.var(pm, nw)[:pconv_dc] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_pconv_dc_$(i)",
         start = 1.0 #comp_start_value(_PM.ref(pm, nw, :convdc, i), "Pdcset", c, 1.0)
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for i in _PM.ids(pm, nw, :convdc)
-            for c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]
-                JuMP.set_lower_bound.(pcdc[i][c], -_PM.ref(pm, nw, :convdc)[i]["Pacrated"][c] * bigM)
-                JuMP.set_upper_bound.(pcdc[i][c], _PM.ref(pm, nw, :convdc)[i]["Pacrated"][c] * bigM)
-            end
+    if bounded # Active power bounds are not defined for the converter terminal related to the metallic return
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            active_pole = first(conductors[i])[1:end-1]
+            JuMP.set_lower_bound.(vars[i][active_pole], -convdc["Pacrated"][active_pole] * bigM)
+            JuMP.set_upper_bound.(vars[i][active_pole], convdc["Pacrated"][active_pole] * bigM)
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :pdc, _PM.ids(pm, nw, :convdc), pcdc)
+    report && sol_component_value_status(pm, nw, :convdc, :pdc, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `pconv_dcg[j]` for `j` in `convdc`"
 function variable_dcside_ground_power(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 1.2 # to account for losses, maximum losses to be derived
-    pcdcg = _PM.var(pm, nw)[:pconv_dcg] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_pconv_dcg_$(i)",
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:pconv_dcg] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_pconv_dcg_$(i)",
         start = 1.0
     ) for i in _PM.ids(pm, nw, :convdc)
     )
+
     if bounded
-        for i in _PM.ids(pm, nw, :convdc)
-            for c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]
-                JuMP.set_lower_bound.(pcdcg[i][c], -_PM.ref(pm, nw, :convdc)[i]["Pacrated"][c] * bigM)
-                JuMP.set_upper_bound.(pcdcg[i][c], _PM.ref(pm, nw, :convdc)[i]["Pacrated"][c] * bigM)
-            end
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound.(vars[i], -convdc["Pacrated"][first(conductors[i])] * bigM)
+            JuMP.set_upper_bound.(vars[i], convdc["Pacrated"][first(conductors[i])] * bigM)
         end
     end
-    report && _PM.sol_component_value(pm, nw, :convdc, :pdcg, _PM.ids(pm, nw, :convdc), pcdcg)
+
+    report && sol_component_value_status(pm, nw, :convdc, :pdcg, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `pconv_dcg_shunt[j]` for `j` in `convdc`"
 function variable_dcside_grounding_shunt_power(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 1.2 # to account for losses, maximum losses to be derived
-    pcdcg_shunt = _PM.var(pm, nw)[:pconv_dcg_shunt] = Dict(i => JuMP.@variable(pm.model,
+    vars = _PM.var(pm, nw)[:pconv_dcg_shunt] = Dict(i => JuMP.@variable(pm.model,
         base_name = "$(nw)_pconv_dcg_shunt_$(i)",
         start = 1.0 #comp_start_value(_PM.ref(pm, nw, :convdc, i), "Pdcset", c, 1.0)
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
-    if bounded
-        for i in _PM.ids(pm, nw, :convdc)
-            JuMP.set_lower_bound.(pcdcg_shunt[i], -_PM.ref(pm, nw, :convdc)[i]["Pacrated"][1] * 0.1 * bigM * _PM.ref(pm, nw, :convdc)[i]["ground_type"])
-            JuMP.set_upper_bound.(pcdcg_shunt[i], _PM.ref(pm, nw, :convdc)[i]["Pacrated"][1] * 0.1 * bigM * _PM.ref(pm, nw, :convdc)[i]["ground_type"])
+    if bounded # The grounding shunt power is bounded by using the limits of the positive (i.e. first) pole
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound(vars[i], -convdc["Pacrated"][1] * 0.1 * bigM * convdc["ground_type"])
+            JuMP.set_upper_bound(vars[i], convdc["Pacrated"][1] * 0.1 * bigM * convdc["ground_type"])
         end
     end
-    report && _PM.sol_component_value(pm, nw, :convdc, :pdcg_shunt, _PM.ids(pm, nw, :convdc), pcdcg_shunt)
+
+    report && _PM.sol_component_value(pm, nw, :convdc, :pdcg_shunt, _PM.ids(pm, nw, :convdc), vars)
 end
 
 "variable: `pconv_dc[j]` for `j` in `convdc`"
 function variable_converter_firing_angle(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
-    phic = _PM.var(pm, nw)[:phiconv] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_phiconv_$(i)",
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:phiconv] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_phiconv_$(i)",
         start = 0
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
     if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(phic[c], 0)
-            JuMP.set_upper_bound.(phic[c], pi)
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound.(vars[i], 0)
+            JuMP.set_upper_bound.(vars[i], pi)
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :phi, _PM.ids(pm, nw, :convdc), phic)
+    report && sol_component_value_status(pm, nw, :convdc, :phi, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `iconv_ac[j]` for `j` in `convdc`"
 function variable_acside_current(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
-    ic = _PM.var(pm, nw)[:iconv_ac] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_iconv_ac_$(i)",
-        start = comp_start_value(_PM.ref(pm, nw, :convdc, i), "P_g", c, 1.0)
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:iconv_ac] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_iconv_ac_$(i)",
+        start = 1.0
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
     if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(ic[c], 0)
-            JuMP.set_upper_bound.(ic[c], convdc["Imax"])
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound.(vars[i], 0)
+            JuMP.set_upper_bound.(vars[i], convdc["Imax"][first(conductors[i])])
         end
     end
 
-    report && _PM.sol_component_value(pm, nw, :convdc, :iconv, _PM.ids(pm, nw, :convdc), ic)
+    report && sol_component_value_status(pm, nw, :convdc, :iconv, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `iconv_ac[j]` and `iconv_ac_sq[j]` for `j` in `convdc`"
@@ -374,37 +377,41 @@ end
 "variable: `vmf[j]` for `j` in `convdc`"
 function variable_converter_filter_voltage_magnitude(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 1.2 # only internal converter voltage is strictly regulated
-    vmf = _PM.var(pm, nw)[:vmf] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_vmf_$(i)",
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:vmf] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_vmf_$(i)",
         start = 0 # start = _PM.ref(pm, nw, :convdc, i, "Vtar")
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
     if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(vmf[c], convdc["Vmmin"] / bigM)
-            JuMP.set_upper_bound.(vmf[c], convdc["Vmmax"] * bigM)
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound.(vars[i], convdc["Vmmin"][first(conductors[i])] / bigM)
+            JuMP.set_upper_bound.(vars[i], convdc["Vmmax"][first(conductors[i])] * bigM)
         end
     end
-    report && _PM.sol_component_value(pm, nw, :convdc, :vmfilt, _PM.ids(pm, nw, :convdc), vmf)
+
+    report && sol_component_value_status(pm, nw, :convdc, :vmfilt, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `vaf[j]` for `j` in `convdc`"
 function variable_converter_filter_voltage_angle(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 2 * pi #
-    vaf = _PM.var(pm, nw)[:vaf] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_vaf_$(i)",
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:vaf] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_vaf_$(i)",
         start = 0
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
     if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(vaf[c], -bigM)
-            JuMP.set_upper_bound.(vaf[c], bigM)
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound.(vars[i], -bigM)
+            JuMP.set_upper_bound.(vars[i], bigM)
         end
     end
-    report && _PM.sol_component_value(pm, nw, :convdc, :vafilt, _PM.ids(pm, nw, :convdc), vaf)
+
+    report && sol_component_value_status(pm, nw, :convdc, :vafilt, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 function variable_converter_internal_voltage(pm::_PM.AbstractPowerModel; kwargs...)
@@ -414,35 +421,39 @@ end
 
 "variable: `vmc[j]` for `j` in `convdc`"
 function variable_converter_internal_voltage_magnitude(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
-    vmc = _PM.var(pm, nw)[:vmc] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_vmc_$(i)",
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:vmc] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_vmc_$(i)",
         start = _PM.ref(pm, nw, :convdc, i, "Vtar")
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
     if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(vmc[c], convdc["Vmmin"])
-            JuMP.set_upper_bound.(vmc[c], convdc["Vmmax"])
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound.(vars[i], convdc["Vmmin"][first(conductors[i])])
+            JuMP.set_upper_bound.(vars[i], convdc["Vmmax"][first(conductors[i])])
         end
     end
-    report && _PM.sol_component_value(pm, nw, :convdc, :vmconv, _PM.ids(pm, nw, :convdc), vmc)
+
+    report && sol_component_value_status(pm, nw, :convdc, :vmconv, _PM.ids(pm, nw, :convdc), conductors, vars)
 end
 
 "variable: `vac[j]` for `j` in `convdc`"
 function variable_converter_internal_voltage_angle(pm::_PM.AbstractPowerModel; nw::Int=_PM.nw_id_default, bounded::Bool=true, report::Bool=true)
     bigM = 2 * pi #
-    vac = _PM.var(pm, nw)[:vac] = Dict(i => JuMP.@variable(pm.model,
-        [c in 1:_PM.ref(pm, nw, :convdc)[i]["conductors"]], base_name = "$(nw)_vac_$(i)",
+    conductors = _PM.ref(pm, nw, :convs_ac_cond)
+    vars = _PM.var(pm, nw)[:vac] = Dict(i => JuMP.@variable(pm.model,
+        [first(conductors[i])], base_name = "$(nw)_vac_$(i)",
         start = 0
     ) for i in _PM.ids(pm, nw, :convdc)
     )
 
     if bounded
-        for (c, convdc) in _PM.ref(pm, nw, :convdc)
-            JuMP.set_lower_bound.(vac[c], -bigM)
-            JuMP.set_upper_bound.(vac[c], bigM)
+        for (i, convdc) in _PM.ref(pm, nw, :convdc)
+            JuMP.set_lower_bound.(vars[i], -bigM)
+            JuMP.set_upper_bound.(vars[i], bigM)
         end
     end
-    report && _PM.sol_component_value(pm, nw, :convdc, :vaconv, _PM.ids(pm, nw, :convdc), vac)
+
+    report && sol_component_value_status(pm, nw, :convdc, :vaconv, _PM.ids(pm, nw, :convdc), conductors, vars)
 end